Consolidate link styling
Ensure that all links use a color transition. Except in a few
circumstances where linkiness isn't a vital characteristic — e.g. tag
lists in post summaries — ensure that there is a non-color change on
hover for accessibility (e.g. underline).

This is technically consolidated but I am not claiming that templates
and hierarchy, nor modularity, are in any way addressed yet.