Turn off ligatures in skills list on CV
Disable ligatures in the skills list on the CV page because there is a
bug with the 'fl' ligature (it renders lowercase) of Google Font's
version of Playfair SC. As noted in the comment, this bug does not
manifest if using `font-variant-caps:all-small-caps` with the "regular"
(not small-caps) Playfair font, so some follow-up is needed.

Also regenerate CV PDF with fix.