Get rid of blank line at beginning of output HTML documents
Prettier will insert a blank line between front-matter and content when
in HTML format. This needs a broader fix, probably, but this front
matter here is not important, so an easy fix is to remove it. This
eliminates leading blank lines throughout the site on HTML documents,
as this is the only HTML template actually used for the whole page.