Add TS to npm scripts; refactor
npm scripts require a refactor to add TypeScript support because there
is a race condition otherwise.

It is necessary to build CSS after eleventy so that content files are
present in `_site` for TailwindCSS to "find." It is not necessary to
"build" TS because the compiled JS output is version-controlled and
already present in `src/_lib` for Eleventy to consume.

This is still feeling more fragile than I'd like. I'm not entirely
convinced we won't have another go-around here.